work on component structure
